+++ src/lib.rs
@@ -44,7 +44,8 @@ mod util;
 
 pub use client::customers::{
     Address, AddressRequest, CreateCustomerRequest, Customer, CustomerId,
-    CustomerPaymentProviderRequest, PaymentProvider, UpdateCustomerRequest,
+    CustomerPaymentProviderRequest, LedgerEntryRequest, LedgerEntry,
+    PaymentProvider, UpdateCustomerRequest, VoidReason,
 };
 pub use client::events::{
     AmendEventRequest, Event, EventPropertyValue, EventSearchParams, IngestEventDebugResponse,