+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning].
 ## [Unreleased] <!-- #release:date -->
 
 * Add support for Plan metadata.
+* Add support for creating customer ledger entries.
 * Bump MSRV to 1.67.
 
 ## [0.6.0] - 2023-05-12